Compute the distribution of differences of replacement samples of two Bernoulli experiments.

Compute the distribution of max(1, nBeffective/nAeffective)*sum(a) - max(1, nAeffective/nBeffective)*sum(b) where a is a 0/1 vector of length nAeffective with each item 1 with independent probability (kA+kB)/(nA+nB), and b is a 0/1 vector of length nBeffective with each item 1 with independent probability (kA+kB)/(nA+nB). The idea is: under this scaling differnces in success rates between the two processes are easilly observed as differences in counts returned by the scaled processes. The method be used to get the exact probability of a given difference under the null hypothesis that both the A and B processes have the same success rate.

Bernoulli_diff_dist(kA, nA, kB, nB, test_rate)

Arguments

kA
number of A successes observed.

nA
number of A experiments.

kB
number of B successes observed.

nB
number of B experiments.

test_rate
numeric, difference in rate of A-B to test. Note: it is best to specify this prior to looking at the data.

Value

Bernoulli difference test statistic.

Details

Note the intent that we are measuring the results of an A/B test with nA %% nB == 0 (no padding needed), or nA >> nB, or nB >> nA. The larger sample is padded so the smaller sample divides evenly into it (allowing faster calculation methods). The sizes of the effective samples are given by nAeffective and nBeffective. The padding will slighly over-estimate confidences due the increased sample size, but if nA and nB are not near each other- this will be a small effect.

Examples

Bernoulli_diff_dist(2000, 5000, 100, 200, 0.1)

#> [1] "Bernoulli difference test: (A=2000/5000=0.4, B=100/200=0.5, prior 0.1 two_sided; p=0.0047)."

Bernoulli_diff_dist(100, 200, 2000, 5000, 0.1)

#> [1] "Bernoulli difference test: (A=100/200=0.5, B=2000/5000=0.4, prior 0.1 two_sided; p=0.0047)."