Compute the distribution of differences of replacement samples of two Bernoulli experiments.

Compute the distribution of max(1, nBeffective/nAeffective)*sum(a) - max*1, nAeffective/nBeffective)*sum(b) where a is a 0/1 vector of length nAeffective with each item 1 with independent probability (kA+kB)/(nA+nB), and b is a 0/1 vector of length nBeffective with each item 1 with independent probability (kA+kB)/(nA+nB). The idea is: under this scaling differnces in success rates between the two processes are easilly observed as differences in counts returned by the scaled processes. The method be used to get the exact probability of a given difference under the null hypothesis that both the A and B processes have the same success rate.

Bernoulli_diff_dist(nA, kA, nB, kB)

Arguments

nA
number of A experiments.

kA
number of A successes observed.

nB
number of B experiments.

kB
number of B successes observed.

Value

experiment summaries

Details

Note the intent that we are measuring the results of an A/B test with nA %% nB == 0 (no padding needed), or nA >> nB, or nB >> nA. The larger sample is padded so the smaller sample divides evenly into it (allowing faster calculation methods). The sizes of the effective samples are given by nAeffective and nBeffective. The padding will slighly over-estimate confidences due the increased sample size, but if nA and nB are not near each other- this will be a small effect.

Examples

Bernoulli_diff_dist(5, 2, 2, 1)

#> $nA
#> [1] 5
#> 
#> $kA
#> [1] 2
#> 
#> $nB
#> [1] 2
#> 
#> $kB
#> [1] 1
#> 
#> $probi
#> [1] 0.4285714
#> 
#> $nAeffective
#> [1] 6
#> 
#> $nBeffective
#> [1] 2
#> 
#> $effective_diff
#> [1] -0.6
#> 
#> $distribution
#>    diff        prob    prob_le    prob_lt     prob_ge     prob_gt
#> 1    -6 0.006394670 0.00639467 0.00000000 1.000000000 0.993605330
#> 2    -5 0.028776015 0.03517068 0.00639467 0.993605330 0.964829315
#> 3    -4 0.053955028 0.08912571 0.03517068 0.964829315 0.910874287
#> 4    -3 0.071007481 0.16013319 0.08912571 0.910874287 0.839866805
#> 5    -2 0.107085743 0.26721894 0.16013319 0.839866805 0.732781062
#> 6    -1 0.152984986 0.42020392 0.26721894 0.732781062 0.579796076
#> 7     0 0.156386491 0.57659041 0.42020392 0.579796076 0.423409585
#> 8     1 0.132089902 0.70868032 0.57659041 0.423409585 0.291319683
#> 9     2 0.120199813 0.82888013 0.70868032 0.291319683 0.171119870
#> 10    3 0.098955020 0.92783515 0.82888013 0.171119870 0.072164850
#> 11    4 0.053955028 0.98179018 0.92783515 0.072164850 0.018209822
#> 12    5 0.016186508 0.99797669 0.98179018 0.018209822 0.002023314
#> 13    6 0.002023314 1.00000000 0.99797669 0.002023314 0.000000000
#>